Build using meson.
Disable dh_autoreconf. It fails because it cannot find gtk-doc.make,
but it is not needed with meson anyway. See also #842246.

Remove libwnck_get_resource@Base symbol. It does not exist anymore
(I believe because meson has a different handling of resources),
but it was not part of the public API anyway, so can be safely dropped.